Cap nhat coed form4

+++ QLKS/DTO/LoaiPhongDTO.cs
@@ -12,7 +12,7 @@ namespace DTO
 
         public string TENLOAIPHONG { get; set; }
 
-        public float    DONGIA { get; set; }
+        public int DONGIA { get; set; }
         
         public int SONGUOI { get; set; }
